more

+++ BinarySearch/50-nthPowOfX.cpp
@@ -5,7 +5,7 @@
  *
  * Creation Date : 22-01-2017
  *
- * Last Modified : Sun Jan 22 20:49:27 2017
+ * Last Modified : Tue Feb 28 21:40:32 2017
  *
  * Created By :  Renne Bai
 **************************************************************************/
@@ -22,7 +22,7 @@ public:
 // Bit manipulation
 class Solution {
 public:
-    double myPow(double x, int n) { 
+    double myPow(double x, long int n) { 
         if(n==0) return 1.0;
         if(n<0) { n = -n; x = 1/x;}
